refactor: Rename hook import to follow consistent naming convention and add new modal hook for enhanced functionality
